% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/correcting.R
\name{correcting}
\alias{correcting}
\alias{correcting,MultiDataSet-method}
\alias{correcting,ExpressionSet-method}
\title{correcting}
\usage{
correcting(
  x,
  reference.c = c("pool", "sample")[1],
  span.n = 1,
  sample_intensity.c = c("median", "mean", "sum")[2],
  title.c = NA,
  col_batch.c = "batch",
  col_injectionOrder.c = "injectionOrder",
  col_sampleType.c = "sampleType",
  figure.c = c("none", "interactive", "myfile.pdf")[2],
  report.c = c("none", "interactive", "myfile.txt")[2]
)

\S4method{correcting}{MultiDataSet}(
  x,
  reference.c = c("pool", "sample")[1],
  span.n = 1,
  sample_intensity.c = c("median", "mean", "sum")[2],
  title.c = NA,
  col_batch.c = "batch",
  col_injectionOrder.c = "injectionOrder",
  col_sampleType.c = "sampleType",
  figure.c = c("none", "interactive", "myfile.pdf")[2],
  report.c = c("none", "interactive", "myfile.txt")[2]
)

\S4method{correcting}{ExpressionSet}(
  x,
  reference.c = c("pool", "sample")[1],
  span.n = 1,
  sample_intensity.c = c("median", "mean", "sum")[2],
  title.c = NA,
  col_batch.c = "batch",
  col_injectionOrder.c = "injectionOrder",
  col_sampleType.c = "sampleType",
  figure.c = c("none", "interactive", "myfile.pdf")[2],
  report.c = c("none", "interactive", "myfile.txt")[2]
)
}
\arguments{
\item{x}{An S4 object of class \code{ExpressionSet} or \code{MultiDataSet}}

\item{reference.c}{Character: sample type to be used as reference for
the correction (as indicated in the 'colnameSampleType' column from the
pData(x); e.g. 'pool')}

\item{span.n}{Numeric: smoothing parameter for the loess regression;
between 0 and 1; (default set to 1)}

\item{sample_intensity.c}{Character: metric to be used when displaying the sample intensities}

\item{title.c}{Character (ExpressionSet): Graphic title: if NA [default] the
title slot from the experimentData will be used}

\item{col_batch.c}{Character: name of the column from pData(x) containing
the batch information (encoded as characters)}

\item{col_injectionOrder.c}{Character: name of the column from pData(x)
containing the injection order information (encoded as numerics)}

\item{col_sampleType.c}{Character:  name of the column from pData(x)
containing the sample type information (encoded as characters)}

\item{figure.c}{Character: File name with '.pdf' extension for the figure;
if 'interactive' (default), figures will be displayed interactively; if 'none',
no figure will be generated}

\item{report.c}{Character: File name with '.txt' extension for the printed
results (call to sink()'); if 'interactive' (default), messages will be
printed on the screen; if 'none', no verbose will be generated}
}
\value{
\code{ExpressionSet} or \code{MultiDataSet} including the corrected
intensities in the expr matrix (matrices)
}
\description{
Signal drift and batch effect correction
}
\examples{
sacurine.eset <- phenomis::reading(system.file("extdata/W4M00001_Sacurine-statistics", package = "phenomis"))
sacurine.eset <- phenomis::correcting(sacurine.eset)
# MultiDataSet (to be done)
}
