% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plots.R
\name{.gg_barplot}
\alias{.gg_barplot}
\title{Barplot with ggplot2}
\usage{
.gg_barplot(
  data.tb,
  x.c = "",
  y.c = "",
  color.c = "",
  title.c = "",
  xlab.c = "",
  ylab.c = "",
  palette.vc = "Set1",
  legend_position.c = c("none", "top")[2],
  geom_text.ls = list(axis.i = 18, bar.i = 10, bar_just.n = 1.5, title.i = 28),
  flip.l = TRUE,
  position_dodge.l = TRUE,
  figure.c = c("interactive", "my_barplot.pdf")[1]
)
}
\arguments{
\item{data.tb}{Tibble (or data frame) containing the information}

\item{x.c}{Character: name of the column with qualitative levels}

\item{y.c}{Character: name of the column with quantitative values}

\item{color.c}{Character: optional name of the column for color information}

\item{title.c}{Character: plot title}

\item{xlab.c}{Character: x label}

\item{ylab.c}{Character: y label}

\item{palette.vc}{Character: either the name of an RColorBrewer palette
(default: 'Set1'; 'Paired' can be useful for parallel plotting) or a vector
manually defining the colors}

\item{legend_position.c}{Character: either 'none', 'bottom', top' (default),
'left', or 'right'}

\item{geom_text.ls}{List of sizes for axis and legend text (default: 18), bar text
(default 10) and adjustment (default 1.5) and title (default: 28)}

\item{flip.l}{Logigal: should the barplot be flipped horizontally (default: TRUE)}

\item{position_dodge.l}{Logical: should the paired bars be displayed side by side
(default: TRUE)}

\item{figure.c}{Character: either 'interactive' for interactive display or
'my_barplot.pdf' for figure saving (only the extension matters)}
}
\value{
invisible ggplot2 object
}
\description{
Barplot with ggplot2
}
\examples{
prometis.mset <- phenomis::reading(system.file("extdata/prometis", package = "phenomis"))
prometis_dims.ls <- Biobase::dims(prometis.mset)
prometis_dims.vn <- unlist(prometis_dims.ls)
prometis_dims.df <- data.frame(omic = rep(names(prometis_dims.ls), each = 2),
dimname = rep(names(prometis_dims.ls[[1]]), times = 2),
count = prometis_dims.vn, stringsAsFactors = FALSE)
phenomis:::.gg_barplot(prometis_dims.df, x.c = "omic", y.c = "count", color.c = "dimname",
title.c = "Prometis", palette.vc = "Paired")
}
