% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plots.R
\name{gg_corplot}
\alias{gg_corplot}
\title{Regression plot with ggplot2}
\usage{
gg_corplot(
  data.tb,
  x.c = "",
  y.c = "",
  color.c = "",
  title.c = NA_character_,
  xlab.c = NA_character_,
  ylab.c = NA_character_,
  label.vc = "",
  theme.c = c("default", "bw", "classic", "dark", "gray", "linedraw", "light", "minimal",
    "void")[2],
  size.ls = list(dot.n = 0.7, lab.i = 20, tick.i = 20, title.i = 20),
  figure.c = c("interactive", "my_corplot.pdf")[1]
)
}
\arguments{
\item{data.tb}{Data frame (or tibble) containing the information}

\item{x.c}{Character: name of the column from the indepedent variable}

\item{y.c}{Character: name of the column from the dependent variable}

\item{color.c}{Character: optional name of the column for color information}

\item{title.c}{Character: plot title}

\item{xlab.c}{Character: x label}

\item{ylab.c}{Character: y label}

\item{label.vc}{Character (vector): either the name of a character column
from the data or a character vector of the same length as the rown number of
the data, containing the feature labeling for outlier display}

\item{theme.c}{character(1): name of the ggplot theme}

\item{size.ls}{List of sizes for dots (default is 0.7), labels
(default is 16), ticks (14) and title (20)}

\item{figure.c}{Character: either 'interactive' for interactive display or
'my_corplot.pdf' for figure saving (only the extension matters)}

\item{palette.vc}{Character: either the name of an RColorBrewer palette
(default: 'Set1'; 'Paired' can be useful for parallel plotting) or a vector
manually defining the colors}
}
\value{
gg plot
}
\description{
Regression plot with ggplot2
}
\examples{
sacurine.se <- reading(system.file("extdata/sacurine", package = "phenomis"))
age.vn <- colData(sacurine.se)[, "age"]
tryptophan.vn <- assay(sacurine.se)["Tryptophan", ]
gg_corplot(data.frame(independent = age.vn, dependent = tryptophan.vn),
  x.c = "independent",
  y.c = "dependent",
  xlab.c = "age",
  ylab.c = "intensity",
  title.c = "Tryptophan"
)
}
