% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenomis-package.R
\docType{package}
\name{phenomis-package}
\alias{phenomis-package}
\alias{phenomis}
\title{Phenomics data analysis and integration}
\description{
Package description
}
\examples{
# See the package vignette

}
\author{
E. A. Thevenot (CEA, LIST, MetaboHUB)

Maintainer: Etienne Thevenot <etienne.thevenot@cea.fr>
}
\keyword{package}
