% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/hypotesting.R
\name{plot_hypotesting}
\alias{plot_hypotesting}
\alias{plot_hypotesting,SummarizedExperiment-method}
\title{Displays the volcano plot and the individual box plots (respectively, dot plots) after a two group statistical test (respectively, a correlation test) has been performed on a \code{SummarizedExperiment} object.}
\usage{
plot_hypotesting(
  x,
  test.c = c("ttest", "limma", "wilcoxon", "pearson", "spearman")[1],
  factor.c,
  adjust.c = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")[5],
  adjust_thresh.n = 0.05,
  title.c = NA,
  class_color.vc = "",
  display_signif.l = TRUE,
  theme.c = c("default", "bw", "classic", "dark", "gray", "linedraw", "light", "minimal",
    "void")[2],
  size.ls = list(class.i = 5, lab.i = 16, point.i = 3, tick.i = 14, title.i = 20),
  figure.c = c("interactive", "interactive_plotly", "myfile.pdf")[2]
)

\S4method{plot_hypotesting}{SummarizedExperiment}(
  x,
  test.c = c("ttest", "limma", "wilcoxon", "pearson", "spearman")[1],
  factor.c,
  adjust.c = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")[5],
  adjust_thresh.n = 0.05,
  title.c = NA_character_,
  class_color.vc = "",
  display_signif.l = TRUE,
  theme.c = c("default", "bw", "classic", "dark", "gray", "linedraw", "light", "minimal",
    "void")[2],
  size.ls = list(class.i = 5, lab.i = 16, point.i = 3, tick.i = 14, title.i = 20),
  figure.c = c("none", "interactive", "interactive_plotly", "myfile.pdf")[2]
)
}
\arguments{
\item{x}{An S4 object of class \code{SummarizedExperiment}}

\item{test.c}{Character: type of test that was used: either 'ttest', 'limma', or 'wilcoxon'}

\item{factor.c}{Character: factor name}

\item{adjust.c}{Character: method for multiple testing correction used in the test}

\item{adjust_thresh.n}{Numeric: significance threshold used}

\item{title.c}{Character: plot title}

\item{class_color.vc}{Character vector: colors of the two compared class labels}

\item{display_signif.l}{Logical: should the individual boxplots of the statistically significant variables be plotted in addition to the main volcano plot? (default: TRUE)}

\item{theme.c}{character(1): name of the ggplot theme}

\item{size.ls}{List of sizes for classes (default: 5), xy labels
(default: 16), points (default: 3), ticks (default: 14) and title
(default: 20)}

\item{figure.c}{Character: either 'interactive' (respectively,
'interactive_plotly') for interactive display with ggplot2 (respectively,
with plotly::ggplotly [default]), or 'my_plot.pdf' (respectively
'my_plot.html') for figure saving (only the extension matters) with
ggplot2 (respectively, with plotly::ggplotly)}
}
\value{
invisible ggplot2 object
}
\description{
Displays the volcano plot and the individual box plots (respectively, dot plots) after a two group statistical test (respectively, a correlation test) has been performed on a \code{SummarizedExperiment} object.
}
\examples{
sacurine.se <- reading(system.file("extdata/sacurine", package = "phenomis"))
sacurine.se <- correcting(sacurine.se, figure.c = "none")
sacurine.se <- sacurine.se[, colData(sacurine.se)[, "sampleType"] != "pool"]
sacurine.se <- transforming(sacurine.se)
sacurine.se <- sacurine.se[, colnames(sacurine.se) != "HU_neg_096_b2"]
sacurine.se <- hypotesting(sacurine.se,
       test.c = "ttest",
       factor_names.vc = "gender",
       figure.c = "none",
       report.c = "none"
)
plot_hypotesting(sacurine.se,
       test.c = "ttest",
       factor.c = "gender",
       adjust.c = "BH",
       adjust_thresh.n = 0.05
)
}
