% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenomis-package.R
\name{search_code}
\alias{search_code}
\title{Searching patterns in code}
\usage{
search_code(
  pattern.c,
  dir.c = getwd(),
  fixed.l = TRUE,
  filename_pattern.c = NULL
)
}
\arguments{
\item{pattern.c}{Character: pattern to be searched in the files}

\item{dir.c}{Character: parent directory of the files}

\item{fixed.l}{Logical: should the pattern be searched 'as is'}

\item{filename_pattern.c}{Character: optional filter of file names to be searched}
}
\value{
nothing is returned
}
\description{
This function is useful for code inspection, development, and debugging
}
\examples{
\dontrun{
phenomis::search_code("aes", filename_pattern.c = ".R")
}
}
