% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/inspecting.R
\name{inspecting}
\alias{inspecting}
\alias{inspecting,MultiDataSet-method}
\alias{inspecting,ExpressionSet-method}
\title{Inspecting}
\usage{
inspecting(
  x,
  pool_as_pool1.l = FALSE,
  pool_cv.n = 0.3,
  span.n = 1,
  sample_intensity.c = c("median", "mean")[1],
  title.c = NA,
  plot_dims.l = TRUE,
  figure.c = c("none", "interactive", "myfile.pdf")[2],
  report.c = c("none", "interactive", "myfile.txt")[2]
)

\S4method{inspecting}{MultiDataSet}(
  x,
  pool_as_pool1.l = FALSE,
  pool_cv.n = 0.3,
  span.n = 1,
  sample_intensity.c = c("median", "mean")[1],
  title.c = NA,
  plot_dims.l = TRUE,
  figure.c = c("none", "interactive", "myfile.pdf")[2],
  report.c = c("none", "interactive", "myfile.txt")[2]
)

\S4method{inspecting}{ExpressionSet}(
  x,
  pool_as_pool1.l = FALSE,
  pool_cv.n = 0.3,
  span.n = 1,
  sample_intensity.c = c("median", "mean")[1],
  title.c = NA,
  plot_dims.l = TRUE,
  figure.c = c("none", "interactive", "myfile.pdf")[2],
  report.c = c("none", "interactive", "myfile.txt")[2]
)
}
\arguments{
\item{x}{An S4 object of class \code{ExpressionSet} or \code{MultiDataSet}}

\item{pool_as_pool1.l}{Logical: should pool be included (as pool1) in the correlation
with the dilution factor?}

\item{pool_cv.n}{Numeric: threshold for the coefficient of variation of the pools}

\item{span.n}{Numeric: span parameter used in the loess trend estimation}

\item{sample_intensity.c}{Character: function to be used to display the global
sample intensity}

\item{title.c}{Character: MultiDataSet: title of the barplot showing the number
of samples and variables in each dataset; ExpressionSet: title of the multipanel
graphic displaying the metrics (if NA -default- the title slot from the experimentData
will be used)}

\item{plot_dims.l}{(MultiDataSet) Logical: should an overview of the number of samples and
variables in all datasets be barplotted?}

\item{figure.c}{Character: File name with '.pdf' extension for the figure;
if 'interactive' (default), figures will be displayed interactively; if 'none',
no figure will be generated}

\item{report.c}{Character: File name with '.txt' extension for the printed
results (call to sink()'); if 'interactive' (default), messages will be
printed on the screen; if 'none', no verbose will be generated}
}
\value{
\code{ExpressionSet} or \code{MultiDataSet} including the computed
in pData and fData sample and variable metrics
}
\description{
Provides numerical metrics and graphical overview of an ExpressionSet or MultiDataSet instance
Please note that all variables with a proportion of missing values > 'max_na_prop.n'
or a variance of 0 will be filtered out at the beginning of the method and
therefore in the output ExpressionSet(s)
}
\examples{
sacurine.eset <- reading(system.file("extdata/sacurine", package = "phenomis"))
sacurine.eset <- inspecting(sacurine.eset)
sacurine.eset <- correcting(sacurine.eset)
sacurine.eset <- inspecting(sacurine.eset)
sacurine.eset <- transforming(sacurine.eset)
sacurine.eset <- inspecting(sacurine.eset)
# MultiDataSet
prometis.mset <- reading(system.file("extdata/prometis", package = "phenomis"))
\dontrun{
prometis.mset <- inspecting(prometis.mset)
}
}
