% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/normalizing.R
\name{normalizing}
\alias{normalizing}
\alias{normalizing,MultiDataSet-method}
\alias{normalizing,ExpressionSet-method}
\title{Normalization of the dataMatrix}
\usage{
normalizing(
  x,
  method.c = "pqn",
  report.c = c("none", "interactive", "myfile.txt")[2]
)

\S4method{normalizing}{MultiDataSet}(
  x,
  method.c = "pqn",
  report.c = c("none", "interactive", "myfile.txt")[2]
)

\S4method{normalizing}{ExpressionSet}(
  x,
  method.c = "pqn",
  report.c = c("none", "interactive", "myfile.txt")[2]
)
}
\arguments{
\item{x}{An S4 object of class \code{ExpressionSet} or \code{MultiDataSet}}

\item{method.c}{Character: Normalization method (default: "pqn")}

\item{report.c}{Character: File name with '.txt' extension for the printed
results (call to sink()'); if 'interactive' (default), messages will be
printed on the screen; if 'none', no verbose will be generated}
}
\value{
\code{ExpressionSet} or \code{MultiDataSet} including the exprs matrix
(list of matrices) with transformed intensities
}
\description{
Normalization of the dataMatrix intensities
}
\examples{
eset <- phenomis::reading(file.path(system.file(package = "phenomis"),
                                    "extdata/W4M00001_Sacurine-statistics"))
eset <- eset[, Biobase::sampleNames(eset) != 'HU_neg_096_b2']
eset <- phenomis::transforming(eset, method.c = "log10")
norm.eset <- phenomis::normalizing(eset, method.c = "pqn")
# MultiDataSet
}
