% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/reading-writing.R
\name{writing}
\alias{writing}
\alias{writing,MultiDataSet-method}
\alias{writing,ExpressionSet-method}
\title{Exporting an ExpressionSet (or MultiDataSet) instance into (subfolders with)
the 3 tabulated files 'dataMatrix.tsv', sampleMetadata.tsv', 'variableMetadata.tsv'}
\usage{
writing(
  x,
  dir.c,
  files.ls = NULL,
  overwrite.l = FALSE,
  report.c = c("none", "interactive", "myfile.txt")[2]
)

\S4method{writing}{MultiDataSet}(
  x,
  dir.c,
  files.ls = NULL,
  overwrite.l = FALSE,
  report.c = c("none", "interactive", "myfile.txt")[2]
)

\S4method{writing}{ExpressionSet}(
  x,
  dir.c,
  files.ls = NULL,
  overwrite.l = FALSE,
  report.c = c("none", "interactive", "myfile.txt")[2]
)
}
\arguments{
\item{x}{An S4 object of class \code{ExpressionSet} or \code{MultiDataSet}}

\item{dir.c}{Character: directory where each dataset should be written}

\item{files.ls}{List: alternatively to the dir.c argument, the full names of the files can be provided as a list}

\item{overwrite.l}{Logical: should existing files be overwritten?}

\item{report.c}{Character: File name with '.txt' extension for the printed
results (call to sink()'); if 'interactive' (default), messages will be
printed on the screen; if 'none', no verbose will be generated}
}
\value{
No object returned.
}
\description{
Note that the \code{dataMatrix} is transposed before
export (e.g., the samples are written column wise in the 'dataMatrix.tsv'
exported file).
}
\examples{
metabo.eset <- phenomis::reading(system.file("extdata/prometis/metabolomics", package="phenomis"))
\dontrun{
writing(metabo.eset, dir.c = file.path(getwd(), "metabolomics"))
}
# MultiDataSet
prometis.mset <- reading(system.file("extdata/prometis",package="phenomis"))
\dontrun{
writing(prometis.mset, dir.c = file.path(getwd(), "prometis"))
# alternatively
writing(prometis.mset,
         dir.c = NA,
         files.ls = list(metabolomics = list(dataMatrix.tsvC = file.path(getwd(), "met_dataMatrix.tsv"),
                                      sampleMetadata.tsvC = file.path(getwd(), "met_sampleMetadata.tsv"),
                                      variableMetadata.tsvC = file.path(getwd(), "met_variableMetadata.tsv")),
                        proteomics = list(dataMatrix.tsvC = file.path(getwd(), "pro_dataMatrix.tsv"),
                                      sampleMetadata.tsvC = file.path(getwd(), "pro_sampleMetadata.tsv"),
                                      variableMetadata.tsvC = file.path(getwd(), "pro_variableMetadata.tsv"))))
}
}
